pyproject.toml
[project]
name = "aiassistentshowcase"
version = "0.1.0"
description = ""
authors = [
    {name = "Lukas Hamm",email = "lukas@lukashamm.dev"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "streamlit (>=1.48.0,<2.0.0)",
    "pydantic-ai-slim[mistral] (>=0.6.2,<0.7.0)",
    "supabase (>=2.18.0,<3.0.0)",
    "tavily-python (>=0.7.10,<0.8.0)",
]

[tool.poetry]
packages = [{include = "aiassistentshowcase", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
